# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Church Management Portal API
  version: 1.0.0
  description: REST API for the Church Management Portal application
  contact:
    name: API Support
    email: support@churchportal.local
servers:
  - url: https://api.churchportal.local/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
paths:
...
  /api/people:
    post:
      summary: Create a new person
      description: Creates a new person in the system
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstName
                - lastName
              properties:
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
                dateOfBirth:
                  type: string
                  format: date
                  example: "1980-01-01"
                gender:
                  type: string
                  enum: [male, female, unspecified]
                  example: unspecified
                email:
                  type: string
                  example: john.doe@example.com
                phone:
                  type: string
                  example: "+1234567890"
      responses:
        '201':
          description: Person created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "123e4567-e89b-12d3-a456-426614174000"
        '400':
          description: Invalid request data

    get:
      summary: List people
      description: Retrieves a list of all people in the system
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of people retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Person'
  /api/people/{id}:
    get:
      summary: Get person by ID
      description: Retrieves a specific person by their ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Person retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '404':
          description: Person not found

    put:
      summary: Update person
      description: Updates an existing person's information
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
                dateOfBirth:
                  type: string
                  format: date
                  example: "1980-01-01"
                gender:
                  type: string
                  enum: [male, female, unspecified]
                  example: unspecified
                email:
                  type: string
                  example: john.doe@example.com
                phone:
                  type: string
                  example: "+1234567890"
      responses:
        '204':
          description: Person updated successfully
        '404':
          description: Person not found
  /api/people/{id}/status:
    post:
      summary: Change person's membership status
      description: Changes the membership status of a person
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  example: active
      responses:
        '204':
          description: Membership status updated successfully
        '404':
          description: Person not found
  /api/families:
    post:
      summary: Create a new family
      description: Creates a new family in the system
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: "Doe Family"
      responses:
        '201':
          description: Family created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "123e4567-e89b-12d3-a456-426614174000"
        '400':
          description: Invalid request data

    get:
      summary: List families
      description: Retrieves a list of all families in the system
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of families retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Family'
  /api/families/{id}:
    get:
      summary: Get family by ID
      description: Retrieves a specific family by their ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Family retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Family'
        '404':
          description: Family not found
  /api/families/{familyId}/members:
    post:
      summary: Add a member to a family
      description: Adds a person as a member of a family
      security:
        - bearerAuth: []
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - personId
              properties:
                personId:
                  type: string
                  example: "123e4567-e89b-12d3-a456-426614174001"
                role:
                  type: string
                  example: member
      responses:
        '204':
          description: Member added successfully
        '404':
          description: Family or person not found
  /api/people/{personId}/relationships:
    get:
      summary: Get person's relationships
      description: Retrieves all relationships for a specific person
      security:
        - bearerAuth: []
      parameters:
        - name: personId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Relationships retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Relationship'
        '404':
          description: Person not found

    post:
      summary: Create a new relationship
      description: Creates a new relationship between two people
      security:
        - bearerAuth: []
      parameters:
        - name: personId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - relatedPersonId
                - type
              properties:
                relatedPersonId:
                  type: string
                  example: "123e4567-e89b-12d3-a456-426614174002"
                type:
                  type: string
                  example: spouse
      responses:
        '201':
          description: Relationship created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "123e4567-e89b-12d3-a456-426614174003"
        '400':
          description: Invalid request data
  /api/relationships/{id}/end:
    post:
      summary: End a relationship
      description: Ends an existing relationship
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                endedAt:
                  type: string
                  format: date-time
                  example: "2023-01-01T00:00:00Z"
      responses:
        '204':
          description: Relationship ended successfully
        '404':
          description: Relationship not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Person:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        dateOfBirth:
          type: string
          format: date
          example: "1980-01-01"
        gender:
          type: string
          enum: [male, female, unspecified]
          example: unspecified
        email:
          type: string
          example: john.doe@example.com
        phone:
          type: string
          example: "+1234567890"
        status:
          type: string
          example: active
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
    Family:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "Doe Family"
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
    Relationship:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        personId:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174001"
        relatedPersonId:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174002"
        type:
          type: string
          example: spouse
        startDate:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        endDate:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
...